backend/pyproject.toml
[project]
name = "ot-tracking-backend"
version = "0.1.0"
description = "Backend for OT Tracking System"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "gspread",
    "bcrypt",
    "pyjwt",
    "pydantic",
    "passlib"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
